agent/requirements.txt
requests>=2.28.0
pywin32>=306
pynput>=1.7.6
Pillow>=9.0.0
pystray>=0.19.4
psutil>=5.9.0
